pyproject.toml
[tool.poetry]
name = "threads-tda"
version = "0.0.1"
description = "Performing Topological Data Analysis on top of the new [Threads](https://www.threads.net/) platform."
authors = ["Zach Bloss <zacharybloss@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "threads_tda"}]

[tool.poetry.dependencies]
python = "^3.10.5"
torch = "2.0.1"
pymongo = {extras = ["srv"], version = "^4.4.1"}
notebook = "^6.5.4"
threads-net = "^0.0.32"
pydantic = "^2.0.2"
python-dotenv = "^1.0.0"
black = "^23.7.0"
kubernetes = "^27.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
